requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.15,<3.16
djangorestframework-simplejwt>=5.2,<5.3
Pillow>=9.5,<10.0
psycopg2-binary>=2.9,<3.0
python-dotenv>=1.0,<2.0

black==23.9.1
flake8==6.0.0
isort==5.12.0
